# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Casilleros ESCOM",
    "version": "1.0.0",
    "description": "Documentación de la API para la gestión de casilleros en ESCOM."
  },
  "servers": [
    { "url": "http://localhost:3001" }
  ],
  "paths": {
...
    "/pagos": {
      "post": {
        "summary": "Registrar pago (subir comprobante de pago)",
        "tags": ["Pagos"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_solicitud": { "type": "integer", "example": 1 },
                  "comprobante": { "type": "string", "format": "binary" }
                },
                "required": ["id_solicitud", "comprobante"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pago registrado correctamente" },
          "400": { "description": "Falta id_solicitud o comprobante" },
          "500": { "description": "Error al registrar pago" }
        }
      }
    },
    "/pagos/{id_pago}/validar": {
      "put": {
        "summary": "Aprobar o rechazar pago (coordinador)",
        "tags": ["Pagos"],
        "parameters": [
          {
            "name": "id_pago",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "validado": { "type": "boolean", "example": true },
                  "estado_pago": { "type": "string", "enum": ["pagado", "no pagado"] }
                },
                "required": ["validado", "estado_pago"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Pago validado correctamente" },
          "400": { "description": "Datos inválidos" },
          "404": { "description": "Pago no encontrado" },
          "500": { "description": "Error al validar pago" }
        }
      }
    },
    "/pagos/solicitud/{id_solicitud}": {
      "get": {
        "summary": "Obtener pagos de una solicitud",
        "tags": ["Pagos"],
        "parameters": [
          {
            "name": "id_solicitud",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de pagos de la solicitud" },
          "500": { "description": "Error al obtener pagos" }
        }
      }
    },
...
    "/asignaciones": {
      "post": {
        "summary": "Crear asignación de casillero (pago validado y casillero disponible)",
        "tags": ["Asignaciones"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_pago": { "type": "integer", "example": 1 },
                  "id_casillero": { "type": "integer", "example": 10 }
                },
                "required": ["id_pago", "id_casillero"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Asignación creada correctamente" },
          "400": { "description": "El pago no está validado o el casillero no está disponible" },
          "500": { "description": "Error al asignar casillero" }
        }
      },
      "get": {
        "summary": "Obtener todas las asignaciones (coordinador)",
        "tags": ["Asignaciones"],
        "responses": {
          "200": { "description": "Lista de asignaciones" },
          "500": { "description": "Error al obtener asignaciones" }
        }
      }
    },
    "/asignaciones/usuario/{id_usuario}": {
      "get": {
        "summary": "Obtener asignaciones de un usuario",
        "tags": ["Asignaciones"],
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de asignaciones del usuario" },
          "500": { "description": "Error al obtener asignaciones" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}